finished version 1

+++ app/Http/Controllers/Api/TodoTasks/V1/TasksController.php
@@ -8,23 +8,42 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use App\Http\Controllers\API\BaseController as BaseController;
+use App\Http\Controllers\API\ApiResponse;
 
-class TasksController extends Controller
+class TasksController extends BaseController
 {
+
 
     public function index()
     {
-        return Task::all();
+        $tasks = Auth::user()->tasks()->get();
+        return $this->sendOkResponse($tasks->toArray(), 'Tasks retrieved successfully.');
     }
 
     public function show($id)
     {
-        return Task::findOrFail($id);
+        $user = Auth::user();
+
+        $task = Task::where('user_id', $user->id)->where('id', $id)->first();
+
+        if (!$task) {
+            return $this->sendBadRequest('Resource cannot be found.');
+        }
+
+        return $this->sendOkResponse($task->toArray(), 'Tasks retrieved successfully.');
     }
 
     public function update(Request $request, $id)
     {
-        $task = Task::findOrFail($id);
+        $user = Auth::user();
+
+        $task = Task::where('user_id', $user->id)->where('id', $id)->first();
+
+        if (!$task) {
+            return $this->sendBadRequest('Resource cannot be found.');
+        }
+
         $task->update(
             [
                 'name' => $request->name,
@@ -32,24 +51,33 @@ class TasksController extends Controller
             ]
         );
 
-        return $task;
+        return $this->sendOkResponse($task->toArray(), 'Task updated successfully.');
     }
 
     public function store(Request $request)
     {
+        $user = Auth::user();
         $task = new Task([
-            'user_id' => 1,//todo
+            'user_id' => $user->id,
             'name' => $request->name,
             'description' => $request->description
         ]);
         $task->save();
-        return $task;
+
+        return $this->sendOkResponse($task->toArray(), 'Task created successfully.');
     }
 
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
+        $user = Auth::user();
+
+        $task = Task::where('user_id', $user->id)->where('id', $id)->first();
+
+        if (!$task) {
+            return $this->sendBadRequest('Resource cannot be found.');
+        }
+
         $task->delete();
-        return '';
+        return $this->sendOkResponse($task->toArray(), 'Task successfully deleted.');
     }
 }
